+++ packages/cydran/package.json
@@ -1,6 +1,6 @@
 {
 	"name": "@cydran/cydran",
-	"version": "0.2.1",
+	"version": "0.2.2",
 	"description": "Unobtrusive Javascript presentation framework",
 	"module": "dist/cydran.js",
 	"main": "./dist/cydran.js",
