--- packages/cydran/src/const/Compare.ts
@@ -1,7 +1,0 @@
-const COMPARE: any = {
-	alpha: (first: string, second: string) => {
-		return (first < second) ? -1 : ((first > second) ? 1 : 0);
-	}
-};
-
-export default COMPARE;
