--- packages/cydran/src/const/Functions.ts
@@ -1,12 +1,0 @@
-function NO_OP_FN() {
-	// Intentionally do nothing
-}
-
-function EMPTY_OBJECT_FN() {
-	return {};
-}
-
-export {
-	NO_OP_FN,
-	EMPTY_OBJECT_FN
-};
