+++ packages/cydran/src/const/InternalPropertyKeys.ts
@@ -1,3 +1,5 @@
+// TODO - Remove this as soon as the factory is refactored to use the registry
+
 enum InternalPropertyKeys {
 	CYDRAN_INTERNAL_FACTORY_DIGESTOR = "cydran.internal.factory.digestor",
 	CYDRAN_INTERNAL_FACTORY_DIGESTION_STATE = "cydran.internal.factory.digestion-state",
