--- packages/cydran/src/const/MimeTypes.ts
@@ -1,5 +1,0 @@
-enum MimeTypes {
-	CYDRAN_REGION = "cydran/region"
-}
-
-export default MimeTypes;
