+++ packages/cydran/src/context/AbstractNamedContextImpl.ts
@@ -9,13 +9,13 @@ import LoggerFactory from "log/LoggerFactory";
 import Initializers from "context/Initializers";
 import InitializersImpl from "context/InitializersImpl";
 import AbstractContextImpl from 'context/AbstractContextImpl';
-import { Context, Stage } from 'context/Context';
+import { Context } from 'context/Context';
 import { VALID_ID } from "const/HardValues";
 import { MutableProperties } from "properties/Property";
 import Registry from "registry/Registry";
 import Scope from "scope/Scope";
 import StageInternalsImpl from "stage/StageInternalsImpl";
-import { argumentsBuilder } from "const/Builder";
+import argumentsBuilder from "function/argumentsBuilder";
 import GlobalContextHolder from 'context/GlobalContextHolder';
 import StageComponent from 'stage/StageComponent';
