+++ packages/cydran/src/function/emptyObject.ts
@@ -1,0 +1,5 @@
+function emptyObject(): Object {
+	return {};
+}
+
+export default emptyObject;
