+++ packages/cydran/src/function/noOp.ts
@@ -1,0 +1,5 @@
+function noOp(): void {
+	// Intentionally do nothing
+}
+
+export default noOp;
