+++ packages/cydran/test/function/emptyObject.spec.ts
@@ -1,0 +1,8 @@
+import emptyObject from "function/emptyObject";
+import { test, expect } from '@jest/globals';
+
+test("emptyObject", () => {
+	const expected: Object = {};
+	const result: Object = emptyObject();
+	expect(result).toEqual(expected);
+});
