+++ packages/integration/package.json
@@ -1,6 +1,6 @@
 {
 	"name": "@cydran/integration",
-	"version": "0.2.1",
+	"version": "0.2.2",
 	"description": "Unobtrusive Javascript presentation framework",
 	"module": "dist/integration.js",
 	"main": "dist/integration.js",
