+++ packages/testsupport/package.json
@@ -1,6 +1,6 @@
 {
 	"name": "@cydran/testsupport",
-	"version": "0.2.1",
+	"version": "0.2.2",
 	"description": "Testing support for Cydran",
 	"module": "dist/testsupport.js",
 	"main": "dist/testsupport.js",
